move TextViewUtils to base-tool module
