+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/viewmodel/CallToActionViewHolder.java
@@ -7,6 +7,7 @@ import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import org.cru.godtools.base.tool.model.view.TextViewUtils;
 import org.cru.godtools.tract.R;
 import org.cru.godtools.tract.R2;
 import org.cru.godtools.tract.util.DrawableUtils;
