+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/viewmodel/CardViewHolder.java
@@ -11,6 +11,7 @@ import android.widget.TextView;
 
 import org.cru.godtools.base.model.Event;
 import org.cru.godtools.base.tool.model.view.ResourceViewUtils;
+import org.cru.godtools.base.tool.model.view.TextViewUtils;
 import org.cru.godtools.tract.R;
 import org.cru.godtools.tract.R2;
 import org.cru.godtools.tract.widget.TractPicassoImageView;
